# benchmarks/locobench_agent/tasks/csharp_blockchain_defi_expert_070_architectural_understanding_expert_01/environment/project/docs/api/rest/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UtilityChain Core Suite – REST API",
    "description": "OpenAPI definition for the UtilityChain Core Suite monolith.  \nThis gateway exposes wallet, staking, governance, NFT and smart–contract features through a unified REST interface.",
    "version": "1.0.0",
    "termsOfService": "https://utilitychain.local/terms",
    "contact": {
      "name": "UtilityChain Engineering",
      "url": "https://utilitychain.local",
      "email": "support@utilitychain.local"
    },
    "license": {
      "name": "Business-Source License 1.1",
      "url": "https://utilitychain.local/license"
    }
  },
  "servers": [
    {
      "url": "https://api.utilitychain.local",
      "description": "Production"
    },
    {
      "url": "https://sandbox.utilitychain.local",
      "description": "Sandbox"
    }
  ],
...
  "paths": {
...
    "/api/v1/wallets": {
      "get": {
        "tags": [ "Wallet" ],
        "summary": "List wallets owned by the authenticated principal",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Paged wallet list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PagedWalletList" }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "Wallet" ],
        "summary": "Create new wallet",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WalletCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Wallet created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Wallet" }
              }
            }
          },
          "409": { "description": "Alias already exists" }
        }
      }
    },
    "/api/v1/wallets/{address}": {
      "get": {
        "tags": [ "Wallet" ],
        "summary": "Get wallet by address",
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          { "$ref": "#/components/parameters/AddressParam" }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Wallet" }
              }
            }
          },
          "404": { "description": "Wallet not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "AddressParam": {
        "name": "address",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[13][a-km-zA-HJ-NP-Z0-9]{25,34}$" },
        "description": "Blockchain wallet address"
      },
...
    },
    "schemas": {
...
      "PagedResultMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "pageSize": { "type": "integer", "example": 25 },
          "total": { "type": "integer", "example": 120 }
        },
        "required": [ "page", "pageSize", "total" ]
      },
      "Wallet": {
        "type": "object",
        "properties": {
          "address": { "type": "string" },
          "alias": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": [ "address", "createdAt" ]
      },
      "WalletCreateRequest": {
        "type": "object",
        "properties": {
          "alias": { "type": "string", "maxLength": 64 }
        }
      },
      "PagedWalletList": {
        "allOf": [
          { "$ref": "#/components/schemas/PagedResultMeta" },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Wallet" }
              }
            },
            "required": [ "items" ]
          }
        ]
      },
...
    }
  }
}